Move subdivision search tests to a separate file
